Memoize expensive introspection.
